commit du 25/07/2025: - Modification configuration du text de jeu

+++ core/src/main/java/com/ctoutweb/aet/core/entity/memoryCardGame/IGameCardData.java
@@ -7,12 +7,12 @@ import com.ctoutweb.aet.core.usecase.generateNewMemoryCardGame.boundary.IGenerat
  * Utilisé pour mettre a jour les données lors de la création d'un jeu
  */
 public interface IGameCardData extends IGenerateNewGameResponse {
-  void setCardToFindQuantity(short quantity);
+  void setCardToFindQuantity(int quantity);
   void setGameTextInformation(IGameTextInformation gameTextInformation);
   void setCardToFindInGame(CardToFind cardToFind);
   void setCards(Card[] cards);
   void setGameLevel(GameLevel gameLevel);
-  void setTimeToObserveBeforeStart(short timeToObserveBeforeStart);
-  void setMaxErrorQuantity(short maxErrorQuantity);
+  void setTimeToObserveBeforeStart(int timeToObserveBeforeStart);
+  void setMaxErrorQuantity(int maxErrorQuantity);
   void setTimeInSecToFinish(int timeInSecToFinish);
 }
